+++ packages/core/src/components/MapContainer.vue
@@ -44,16 +44,21 @@
 </template>
 
 <script lang="ts">
-import Vue from 'vue'
+import Vue, { PropType } from 'vue'
 import { mapActions, mapGetters, mapMutations } from 'vuex'
 import api from '@masterportal/masterportalapi/src/maps/api'
 import { MoveHandle } from '@polar/components'
 import Hammer from 'hammerjs'
 import i18next from 'i18next'
 import { defaults } from 'ol/interaction'
-import { LanguageOption, MoveHandleProperties } from '@polar/lib-custom-types'
+import {
+  LanguageOption,
+  MapConfig,
+  MoveHandleProperties,
+} from '@polar/lib-custom-types'
 import { SMALL_DISPLAY_HEIGHT, SMALL_DISPLAY_WIDTH } from '../utils/constants'
 import { addClusterStyle } from '../utils/addClusterStyle'
+import { setupStyling } from '../utils/setupStyling'
 import MapUi from './MapUi.vue'
 // NOTE: OpenLayers styles need to be imported as the map resides in the shadow DOM
 import 'ol/ol.css'
@@ -67,7 +72,7 @@ export default Vue.extend({
   },
   props: {
     mapConfiguration: {
-      type: Object,
+      type: Object as PropType<MapConfig>,
       required: true,
     },
   },
@@ -143,7 +148,7 @@ export default Vue.extend({
         },
       }
     )
-
+    setupStyling(this.mapConfiguration, map)
     this.setMap(map)
     this.updateDragAndZoomInteractions()
     if (this.mapConfiguration.extendedMasterportalapiMarkers) {