+++ packages/plugins/Pins/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## unpublished
+## 2.0.0
 
 - Breaking: Upgrade peerDependency `ol` from `^7.1.0` to `^9.2.4`.
 - Breaking: Remove support of configuring `movable` with a `boolean`. Please see the documentation for more information on how to configure this parameter.